--- src/shared/ui/Dropdown/index.ts
@@ -1,1 +1,0 @@
-export { Dropdown } from './ui/Dropdown';
