--- src/shared/ui/Listbox/index.ts
@@ -1,1 +1,0 @@
-export { Listbox, type ListboxItem } from './ui/Listbox/Listbox';
