+++ src/shared/ui/Popups/index.ts
@@ -1,0 +1,3 @@
+export { Listbox, type ListboxItem } from './ui/Listbox/ui/Listbox/Listbox';
+
+export { Dropdown } from './ui/Dropdown/ui/Dropdown';
